js: more refactoring and cleaning up
