+++ js/src/magika-prediction.ts
@@ -1,6 +1,6 @@
 import { ContentTypeInfo } from "./content-type-info";
 import { ContentTypeLabel } from "./content-type-label";
-import { OverwriteReason } from "./overwrite_reason";
+import { OverwriteReason } from "./overwrite-reason";
 
 export interface MagikaPrediction {
   dl: ContentTypeInfo;
