+++ js/src/model-prediction.ts
@@ -1,0 +1,8 @@
+import { DataTypeMap, NumericDataType } from "@tensorflow/tfjs";
+import { ContentTypeLabel } from "./content-type-label.js";
+
+export interface ModelPrediction {
+  label: ContentTypeLabel;
+  score: number;
+  scores: DataTypeMap[NumericDataType];
+}
